refactor: 로그아웃 탭 모달로 변경
