+++ src/app/mypage/_components/AccountSettingTabs.tsx
@@ -1,20 +1,16 @@
 import ModalItem from "@/components/ui/Modal";
-import useMediaQuery from "@/lib/hooks/common/useMediaQuery";
-import { useAuthStore } from "@/lib/store/authStore";
 import { CaretRight } from "@phosphor-icons/react";
 import { useState } from "react";
 import AccountDeleteContent from "./AccountDeleteContent";
 import AgreementContent from "./AgreementContent";
+import LogOutContent from "./LogOutContent";
 import ServiceContent from "./ServiceContent";
-import SignoutButton from "./SignoutButton";
 
-type AccountModalMode = "agreement" | "serviceRule" | "deleteAccount" | "mobile";
+type AccountModalMode = "agreement" | "serviceRule" | "deleteAccount" | "logout";
 
 const AccountSettingTabs = () => {
-  const { deleteUser } = useAuthStore();
   const [modal, setModal] = useState(false);
   const [mode, setMode] = useState<AccountModalMode>("agreement");
-  const isTabletOrSmaller = useMediaQuery("(max-width: 768px)");
   const menuItems: { title: string; mode: AccountModalMode }[] = [
     {
       title: "서비스 약관",
@@ -25,31 +21,19 @@ const AccountSettingTabs = () => {
       mode: "agreement"
     },
     {
-      title: isTabletOrSmaller ? "PC 버전" : "모바일 버전",
-      mode: "mobile"
+      title: "로그아웃",
+      mode: "logout"
     },
     { title: "탈퇴하기", mode: "deleteAccount" }
   ];
 
   const openModal = (mode: AccountModalMode) => {
-    if (mode !== "mobile") {
-      setModal(true);
-      setMode(mode);
-    } else {
-      handleClick();
-    }
+    setModal(true);
+    setMode(mode);
   };
 
   const closeModal = () => {
     setModal(false);
-  };
-
-  const handleDeleteAccount = async () => {
-    deleteUser();
-  };
-
-  const handleClick = () => {
-    alert("서비스 준비 중입니다.");
   };
 
   return (
@@ -65,16 +49,14 @@ const AccountSettingTabs = () => {
             <CaretRight className="text-title1 tb:text-body" />
           </button>
         ))}
-        <SignoutButton />
       </div>
 
       {/* 리팩터링이 하고 싶은데... 아이디어가 없다... */}
       <ModalItem isOpen={modal} onClose={closeModal} className="mb:p-4">
         {mode === "serviceRule" && <ServiceContent closeModal={closeModal} />}
-        {mode === "deleteAccount" && (
-          <AccountDeleteContent closeModal={closeModal} handleDeleteAccount={handleDeleteAccount} />
-        )}
+        {mode === "deleteAccount" && <AccountDeleteContent closeModal={closeModal} />}
         {mode === "agreement" && <AgreementContent closeModal={closeModal} />}
+        {mode === "logout" && <LogOutContent closeModal={closeModal} />}
       </ModalItem>
     </>
   );
